% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_sales_ts}
\alias{clean_sales_ts}
\title{Czyszczenie danych sprzedażowych}
\usage{
clean_sales_ts(data, aggregate_by = "day")
}
\arguments{
\item{data}{Ramka danych.}

\item{aggregate_by}{Okres agregacji: "day", "week", "month".}
}
\description{
Usuwa braki, agreguje sprzedaż, promocje oraz flagi świąteczne.
}
