% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{compute_sales_metrics}
\alias{compute_sales_metrics}
\title{Obliczanie metryk biznesowych}
\usage{
compute_sales_metrics(data)
}
\arguments{
\item{data}{Oczyszczona ramka danych.}
}
\description{
Oblicza m.in. sprzedaż całkowitą, średnią, zmienność, wpływ promocji oraz odstępy między szczytami.
}
