% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_management_summary}
\alias{create_management_summary}
\title{Tworzenie podsumowania biznesowego}
\usage{
create_management_summary(data)
}
\arguments{
\item{data}{Ramka danych.}
}
\description{
Generuje raport, identyfikując ekstrema, TOP/BOTTOM sklepów oraz kategorii.
}
