% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_prognosis}
\alias{create_prognosis}
\title{Prognozowanie szeregów czasowych}
\usage{
create_prognosis(data, periods = 30, method = "prophet", zoom_days = 150)
}
\arguments{
\item{data}{Oczyszczona ramka danych.}

\item{periods}{Liczba okresów do przodu.}

\item{method}{Wybór metody: "arima" lub "prophet".}

\item{zoom_days}{Liczba ostatnich dni do wyświetlenia na wykresie.}
}
\description{
Generuje prognozy (Prophet z uwzględnieniem świąt i ARIMA).
}
