% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{load_sales_data}
\alias{load_sales_data}
\title{Wczytywanie danych sprzedażowych}
\usage{
load_sales_data(train_path, stores_path = NULL, holidays_path = NULL)
}
\arguments{
\item{train_path}{Ścieżka do pliku z danymi treningowymi.}

\item{stores_path}{Ścieżka do pliku z metadanymi sklepów (opcjonalnie).}

\item{holidays_path}{Ścieżka do pliku z kalendarzem świąt (opcjonalnie).}
}
\value{
Ramka danych (tibble).
}
\description{
Wczytuje pliki CSV i opcjonalnie łączy je (sklepy, święta) w jedną ramkę danych.
}
