% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_category_trends}
\alias{plot_category_trends}
\title{Wizualizacja trendów dla najlepszych kategorii}
\usage{
plot_category_trends(data)
}
\arguments{
\item{data}{Oczyszczona ramka danych.}
}
\description{
Generuje wykres trendów (product trends) dla 5 najpopularniejszych kategorii w czasie.
}
