% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_promotion_impact}
\alias{plot_promotion_impact}
\title{Analiza skuteczności promocji}
\usage{
plot_promotion_impact(data)
}
\arguments{
\item{data}{Oczyszczona ramka danych.}
}
\description{
Bada korelację między akcjami promocyjnymi a wygenerowanym wolumenem.
}
