% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_sales_trends}
\alias{plot_sales_trends}
\title{Wizualizacja trendów sprzedażowych}
\usage{
plot_sales_trends(data)
}
\arguments{
\item{data}{Ramka danych.}
}
\description{
Generuje estetyczny wykres liniowy trendu wolumenu.
}
