% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_top_rankings}
\alias{plot_top_rankings}
\title{Wizualizacja rankingów biznesowych}
\usage{
plot_top_rankings(summary_data)
}
\arguments{
\item{summary_data}{Lista wygenerowana przez funkcję create_management_summary().}
}
\description{
Generuje wykresy słupkowe dla TOP 5 sklepów i kategorii na podstawie podsumowania.
}
