% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sales_ts_logic}
\alias{sales_ts_logic}
\title{Funkcja wyższego rzędu do logiki biznesowej}
\usage{
sales_ts_logic(data, FUN, store_id = NULL, category = NULL)
}
\arguments{
\item{data}{Ramka danych.}

\item{FUN}{Funkcja do wykonania (np. compute_sales_metrics).}

\item{store_id}{Wybrany numer sklepu.}

\item{category}{Wybrana kategoria (family).}
}
\description{
Umożliwia zaaplikowanie innej funkcji (np. plot, metrics) na wyfiltrowanym podzbiorze danych.
}
