% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{validate_sales_ts}
\alias{validate_sales_ts}
\title{Walidacja danych szeregów czasowych}
\usage{
validate_sales_ts(data)
}
\arguments{
\item{data}{Ramka danych.}
}
\value{
Wypisuje raport walidacyjny w konsoli i zwraca wejściową ramkę.
}
\description{
Sprawdza jakość danych: braki, duplikaty, poprawność dat oraz luki w szeregu.
}
